Move isValidGitBranchName to a separate file
