+++ apps/webapp/app/services/upsertBranch.server.ts
@@ -5,6 +5,7 @@ import { createApiKeyForEnv, createPkApiKeyForEnv } from "~/models/api-key.serve
 import { logger } from "./logger.server";
 import { getLimit } from "./platform.v3.server";
 import { type CreateBranchOptions } from "~/routes/_app.orgs.$organizationSlug.projects.$projectParam.env.$envParam.branches/route";
+import { isValidGitBranchName } from "~/v3/validGitBranch";
 
 export class UpsertBranchService {
   #prismaClient: PrismaClient;
@@ -165,37 +166,6 @@ export async function checkBranchLimit(
     limit,
     isAtLimit: used >= limit,
   };
-}
-
-export function isValidGitBranchName(branch: string): boolean {
-  // Must not be empty
-  if (!branch) return false;
-
-  // Disallowed characters: space, ~, ^, :, ?, *, [, \
-  if (/[ \~\^:\?\*\[\\]/.test(branch)) return false;
-
-  // Disallow ASCII control characters (0-31) and DEL (127)
-  for (let i = 0; i < branch.length; i++) {
-    const code = branch.charCodeAt(i);
-    if ((code >= 0 && code <= 31) || code === 127) return false;
-  }
-
-  // Cannot start or end with a slash
-  if (branch.startsWith("/") || branch.endsWith("/")) return false;
-
-  // Cannot have consecutive slashes
-  if (branch.includes("//")) return false;
-
-  // Cannot contain '..'
-  if (branch.includes("..")) return false;
-
-  // Cannot contain '@{'
-  if (branch.includes("@{")) return false;
-
-  // Cannot end with '.lock'
-  if (branch.endsWith(".lock")) return false;
-
-  return true;
 }
 
 export function sanitizeBranchName(ref: string): string | null {
